Implementato mini Dizionario WN
