+++ src/main/java/com/github/bot/curiosone/core/nlp/tokenizer/PosT.java
